# server/api/openapi.yaml
openapi: 3.0.3
info:
  title: AREA
  description: Area for epitech
  version: 1.0.0
servers:
- url: /
...
paths:
...
  /user:
    put:
      tags:
      - user
      summary: Update user
      description: This can only be done by the logged in user.
      operationId: updateUser
      requestBody:
          content:
              application/json:
                  schema:
                      $ref: "#/components/schemas/User"
      responses:
        "400":
          description: Invalid user supplied
        "404":
          description: User not found
      x-swagger-router-controller: User
    post:
      tags:
      - user
      summary: Create user
      description: This can only be done once at the registration page
      operationId: createUser
      requestBody:
          content:
              application/json:
                  schema:
                      $ref: "#/components/schemas/User"
      responses:
        "200":
          description: "successful operation, user is logged in"
        "400":
          description: Invalid email supplied
        "413":
          description: Invalid password supplied
      x-swagger-router-controller: User
    delete:
      tags:
      - user
      summary: Delete user
      description: This can only be done by the logged in user. Delete user from database
        and log him out
      operationId: deleteUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string                
      responses:
        "200":
          description: User has been deleted
        "400":
          description: User not logged in
      x-swagger-router-controller: User
...
components:
  schemas:
    User:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
          format: password
      xml:
        name: User
...